github-runner-authorizer/requirements.txt
cryptography==3.0
Everett==1.0.2
kubernetes==11.0.0
PyJWT==1.7.1
requests==2.24.0
